feat (usuario) : :sparkles: se cambio el modelo del usuario
